refactor: extract IssueScreenshotAnnotationTypes.swift from IssueScreenshotAnnotationRenderer.swift
Extracts 2 support structs (RenderedIssueScreenshotAsset,
AnnotatedScreenshotExport) from IssueScreenshotAnnotationRenderer.swift into
a new IssueScreenshotAnnotationTypes.swift.

Byte-identical move; SHA verified against origin/main.

Reduces IssueScreenshotAnnotationRenderer.swift from 197 to 179 lines.

Tests: 667 (unchanged).
